migrations refactoring and commentLike migration added

+++ src/main/resources/db/migration/V5__create_comment.sql
@@ -3,7 +3,7 @@ create table comment (
     author_id int8,
     comment_text text,
     is_blocked boolean not null,
-    parent_id int,
+    parent_id bigserial,
     time timestamp not null,
-    post_id int8 not null,
+    post_id bigserial not null,
     primary key (id));
