+++ src/main/resources/db/migration/V6__create_post_like.sql
@@ -1,6 +1,6 @@
 create table post_like (
     id  bigserial not null,
     time timestamp not null,
-    person_id int8 not null,
-    post_id int8 not null,
+    person_id bigserial not null,
+    post_id bigserial not null,
     primary key (id));
